SubmitAnswerDialog refactored - not in use yet
